+++ app/calendar/caldav/caldav-service.server.test.ts
@@ -4,7 +4,7 @@ import {
   verifyConnectionDetails,
   createCalDavEvent,
   formatDateAsICS,
-} from "."
+} from "./caldav-service"
 import { GenericContainer, StartedTestContainer } from "testcontainers"
 import * as path from "path"
 import { addMinutes } from "date-fns"
